+++ src/Trait.mjs
@@ -1,13 +1,14 @@
 import { typeName } from './Data.mjs';
 
 export const isTrait = Symbol('isTrait'),
+    all = Symbol('all'),
     apply = Symbol('apply');
 
 const protoTrait = () => { }
 protoTrait[isTrait] = true;
 protoTrait[apply] = function (instance, ...args) {
     const name = instance[typeName],
-        fn = this[name];
+        fn = this[name] ?? this[all];
     if (!fn)
         throw new TypeError(`no trait defined for ${name}`)
     return fn.call(this, instance, ...args)